% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_weights.R
\name{generate_weights}
\alias{generate_weights}
\title{Compute all intersection hypotheses and corresponding weights for a given graph}
\usage{
generate_weights(G, w)
}
\arguments{
\item{G}{Graph represented as transition matrix}

\item{w}{A numeric vector of weights}
}
\description{
Compute all intersection hypotheses and corresponding weights for a given graph
}
