% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{permutations}
\alias{permutations}
\title{Generate all permutations for a set of length n}
\usage{
permutations(n)
}
\arguments{
\item{n}{set length}
}
\description{
Generate all permutations for a set of length n
}
