% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_weights.R
\name{print_weights}
\alias{print_weights}
\title{Print weights in human-readable format}
\usage{
print_weights(wts)
}
\arguments{
\item{wts}{data.frame of output from /code{generate_weights}}
}
\description{
Print weights in human-readable format
}
