% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_edges.R
\name{update_edges}
\alias{update_edges}
\title{Recursively compute edge weights for a given graph and intersection hypothesis}
\usage{
update_edges(h, G, w)
}
\arguments{
\item{h}{A numeric vector with only binary entries (0,1).}

\item{G}{Graph represented as transition matrix.}

\item{w}{A numeric vector of weights.}
}
\value{
A weight vector
}
\description{
Recursively compute edge weights for a given graph and intersection hypothesis
}
